Add a ConsistentHashingRouter
Summary:
When asked for a DB, the ConsistentHashingRouter will look up in a consistent
hash for the requested key.  The ConsistentHashingRouter has no notion of
connection status or availability, so it may/will give back DB numbers that are
not available.

If a client receives a DB number back from ConsistentHashingRouter and that DB
is unavailable for some reason, the client can re-request a db_number from the
router and pass a retry_for= kwarg that tells the router that the db num
provided in the retry_for arg was unavailable and should be marked as down.
ConsistentHashingRouter will then take the server out of the hash and return a
new db_num as a response.  If no hosts are online, the router will raise a
HostListExhaused exception.

To get hosts **back** in to the hash, the router will flush all down connections
at some interval based on the number of requests to get_db().  When that
threshold is crossed (currently, 100,000), the router flushes all its
connections that are marked as down and then returns the requested host.  That
host may still be down, and if so the client may mark it down again with the
retry_for kwarg.

There were a few other implementations on ways to track host availability that
were rejected:

1. Have the client manage host availability by telling the router to remove or
add servers.  This was deemed to complicated of an interface for clients, since
the notion of adding or removing hosts is unique to this router.
2. Have the router check on host availability every request, by letting clients
define a callable that the router would call when get_db() was called, to
guarantee a working host every call to get_db().  The overhead of checking
connectivity every request was deemed as too high.
3. Have the router check host availability through the callable (from #2 above)
"every so often."  This was the most viable rejected option, but was decided
that having the router  do more than just routing, in this case checking
connectivity, increased coupling too much.  Plus, since the router will still
give back DB numbers that may not be connected, the client still has to do the
work of understanding connection errors and taking the appropriate action - so
it's not a huge win.

Maniphest T117

Test Plan: Check out the newly added tests and also run the whole test suite.

Reviewers: dcramer, tail

Reviewed By: dcramer

CC: jeff, dcramer

Differential Revision: http://phabricator.local.disqus.net/D395